Made it so tester links correctly, fixed minor bugs

+++ src/private/coredata.h
@@ -15,7 +15,7 @@ typedef struct {
 		SBCell* frontbuff;
 		
 		SBCell* backbuff;
-		int backbuffOffset;
+		size_t backbuffOffset;
 	} Buffers;
 
 	struct {
